requirements.txt
flask==3.0.2
flask-login==0.6.3
werkzeug==3.0.1
flask-socketio==5.3.6
python-socketio==5.11.1
pandas==2.2.3
numpy==2.2.5
SpeechRecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
wikipedia-api==0.6.0
google-api-python-client==2.118.0
google-generativeai==0.3.2
PyAudio==0.2.14
pydub==0.25.1
--only-binary :all: 